Changes for Device
something is different between device and simulator

+++ BFRecordingPathHelper.h
@@ -17,7 +17,7 @@
 -(void)resetPathHelper;
 
 #pragma mark -- for recording
--(int)getTotalRecordNumber;
+-(NSInteger)getTotalRecordNumber;
 -(NSURL*)getNextRecordingPath;
 -(NSString*)getDirPath;
 -(NSArray*)getEnumableRecordingPath;
